- Added unit tests for Teacher.cpp and Utils.cpp and added coverage report - Now using 32b (unsigned int) instead of 16b (unsigned short int) for students IDs to avoid truncation. - Now using external IDs instead of vectors indices for parsing. - Now using config.cpp file to define GLOBAL_CONFIG variable only once. - Updated README file about tests and test coverage - Reworked Makefile - Deleted useless simulate run shell files

+++ include/Student.h
@@ -7,7 +7,7 @@ using std::string;
 
 struct Student
 {
-    unsigned short int mId{};
+    unsigned int mId{};
     string mName;
     bool mHasAccommodations{};
     unsigned short int mEffectivePresentationLength{};
